added option to add and delete exercises that will be saved with SwiftData
