+++ FitFlow.xcodeproj/xcuserdata/margaritamayer.xcuserdatad/xcdebugger/Breakpoints_v2.xcbkptlist
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="UTF-8"?>
+<Bucket
+   uuid = "4D04BCE9-58B8-460C-A9B6-1784DD0896FB"
+   type = "1"
+   version = "2.0">
+</Bucket>
